Fixes for kodi submission
- Licence and fanart update
- Added 2-letter language codes in ll_CC format, e.g. en_GB.
- Removed unncessary headers e.g. #!/usr/bin/python
- Removal of extraneous files
- Removed __pychache__ files
- Fixes for Kodi Addon Checker fail
- Fix for error triggered when clean_filenames function receives unescaped strings with commas
