--- script.copacetic.helper/.gitattributes
@@ -1,21 +1,0 @@
-# Auto detect text files and perform LF normalization
-* text=auto
-* text eol=lf
-*.png binary
-*.jpg binary
-*.ttf binary
-
-# Custom for Visual Studio
-*.cs     diff=csharp
-
-# Standard to msysgit
-*.doc	 diff=astextplain
-*.DOC	 diff=astextplain
-*.docx diff=astextplain
-*.DOCX diff=astextplain
-*.dot  diff=astextplain
-*.DOT  diff=astextplain
-*.pdf  diff=astextplain
-*.PDF	 diff=astextplain
-*.rtf	 diff=astextplain
-*.RTF	 diff=astextplain
