+++ script.copacetic.helper/resources/language/resource.language.en_gb/strings.po
