+++ script.copacetic.helper/resources/lib/plugin/json_map.py
@@ -1,5 +1,3 @@
-#!/usr/bin/python
-#coding: utf-8
 # author: realcopacetic, sualfred
 
 JSON_MAP = {
