+++ script.copacetic.helper/resources/lib/plugin/library.py
@@ -1,5 +1,3 @@
-#!/usr/bin/python
-# coding: utf-8
 # author: realcopacetic, sualfred
 
 import xbmc
