+++ script.copacetic.helper/resources/lib/service/monitor.py
@@ -1,5 +1,3 @@
-#!/usr/bin/python
-# coding: utf-8
 # author: realcopacetic
 
 import xml.etree.ElementTree as ET
