+++ script.copacetic.helper/resources/lib/service/player.py
@@ -1,5 +1,3 @@
-#!/usr/bin/python
-# coding: utf-8
 # author: realcopacetic
 
 from xbmc import Player
