+++ script.copacetic.helper/resources/lib/utilities.py
@@ -1,5 +1,3 @@
-#!/usr/bin/python
-# coding: utf-8
 # author: realcopacetic
 
 import json
