+++ script.copacetic.helper/script.py
@@ -1,5 +1,3 @@
-#!/usr/bin/python
-# coding: utf-8
 # author: realcopacetic
 
 from resources.lib.script.actions import *
